split track and event
